refactor e2e tests setup
