+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from '@src/constants'
 export * from '@lib/errors'
 export * from '@lib/util'
+export * from '@src/operations'
 
 export { Registry as Services } from '@services/registry'
-export { Operation } from '@src/operations'
 export { getSchema, getValidData as validate } from '@src/validation'
