+++ tsconfig.build.json
@@ -7,6 +7,7 @@
     "tsconfig.json",
     "jest.setup.ts",
     "jest.config.ts",
+    "jest.config.e2e.ts",
     "dist",
     "node_modules",
     "**/__tests__/**/*",
