+++ tsconfig.json
@@ -62,6 +62,7 @@
     "**/__tests__/**/*",
     "tsconfig.json",
     "jest.setup.ts",
+    "jest.config.e2e.ts",
     "jest.config.ts",
     "node_modules/cdktf/lib/testing/adapters/jest.d.ts"
   ],
